Add crud controllers? try to config travis for docker

+++ main-service/src/main/java/com/dyvak/main/model/service/LogisticService.java
@@ -12,7 +12,7 @@ public interface LogisticService<T> {
 
     void create(T dto);
 
-    void update(Long id, T dto);
+    void update(int id, T dto);
 
     void delete(Long[] ids);
